Add `inferno` cache and switch CI to build with Nix (#24)
Closes #21 

There are some instructions in the readme on how to enable the new
Inferno cache for local development.

CI will now build with Nix and push the artifacts to the cache. If you
configure the cache locally you can pull them, making most Nix rebuilds
much faster.

CI jobs will generally be faster. They'll only be slower (albeit
considerably slower) when large updates happen with dependencies (e.g. a
new haskell.nix revision)

I also disabled HLS in the dev shells (this is unrelated to the changes
in this PR but I discovered it while making the changes here; see #25
for more details)