LSP: error if script evaluates to a function (#48)
I added a check in the LSP that checks if the script type returns a
function (by checking if it expects more inputs than input parameters
supplied). This should warn users and make it less likely that they
confuse the input parameters that are selected in the frontend UI with
the parameters of a lambda function.