Add array indexing and pattern matching (#55)
This PR adds support for array indexing `[1, 2, 3] !! 2 == Array.get [1,
2, 3] 2 == 3` and array pattern matching:
```
fun a -> match a with { | [] -> 0 | [x] -> 1 | [_, y] -> 2 | _ -> 3 }
```

Currently we only allow literal patterns like `[x, _, ...]`, but no
cons-style patterns like Haskell's `x : xs`. This means you must have a
final `| _ -> ...` or `| a -> ...` pattern in order for the pattern
match to be exhaustive.