Fix `torch` versions, document, use on CI (#67)
Closes #65. Previously `devShells..pytorch` was using an incorrect
version (one major version too high). I added something to the docs
mentioning `devShells..pytorch` and also used it to run the MNIST
example on CI. This way we will always be using the same version
everywhere. I temporarily modified the workflow file in order to ignore
the cache and ran the CI job manually and it worked (otherwise that part
wouldn't run because it's been cached)