[inferno-vc] Simplified InfernoVCOperations typeclass usage (#77)
Also removed `getAllHeads` and added `deleteAutosavedVCObjectsOlderThan`
instead to allow for more efficient implementations.

The changes to the class that make the usage simpler are:

- Removed the `De/Serializable` associated constraints which were
applied on the `group` and `author` type variables of `VCMeta` and added
a `Group` and `Author` associated types instead. See the changes in
`runServer` and `runServerConfig` call sites.
This change is not only cosmetic, it also gives knowledge of `author`
and `group` to all type-class method which can use it to deserialise
stuff without resorting to AmbiguousTypes or Proxies (which moves the
burden of keeping track of the type witnesses to the call-sites)

- Made the constraint `AsType err VCStoreError` that `runServer/Config`
requires explicit on the type-class. This is so implementors don't
assume they can ignore it until they're faced with it when attempting to
make the call-sites compile.


Removing `getAllHeads` is motivated by the fact that the only caller was
`Inferno.VersionControl.deleteStaleAutosavedVCObjects`. This function
used an implementation that can be made much more efficient if the
backend knows that what is actually meant is "give me all the heads _so
I can purge some auto-saved objects_" so it has been moved to the
typeclass in its place