Add `inferno-ml-server-types` package (#102)
Adds the `inferno-ml-server-types` package containing the API for a
server that can evaluate Inferno scripts using ML primitives. There is
no actual implementation of `inferno-ml-server` here

This and the whole Inferno ML infrastructure is still a WIP. However, a
WIP implementation will need to be merged here in order to merge some
WIPs in other repos. The following at least are not yet complete:
- making a polymorphic prelude for `inferno-ml` that doesn't depend on
`hasktorch` (helpful for typechecking in cases where we can't use
`hasktorch`)
- golden and other tests for the new types; however, since many of the
types are polymorphic on something, it probably makes more sense to just
add them in one of the places where `inferno-ml-server` or the bridge
server are implemented
- a metrics endpoint for `inferno-ml-server`

I've also made a few changes to `inferno-ml`, which are listed in the
changelog

---------

Co-authored-by: Siddharth Krishna <siddharth-krishna@users.noreply.github.com>