[inferno-ml-server] Add `terminated` to DB types (#113)
This adds a `terminated` column to most of the DB tables, corresponding
to a `Maybe UTCTime` in the Haskell types. This lets us track when DB
types have been "deleted":

- a `Just time` means that the DB row was "deleted" at `time`;
`inferno-ml-server` will not use these rows for inference evaluation or
model caching
- `Nothing` means that the field is still active

Originally I was going to use a Postgres enum of `active`/`passive`. But
then we would need another column to record the time, which would be
messy and redundant, since the semantics of a nullable `timestamptz` are
essentially the same