[inferno-ml] Map model idents to IDs in script metadata (#128)
Two large changes:

- The `ScriptMetadata` now contains the model IDs, which are mapped to
their identifiers. The `InferenceParam` type no longer contains the
models. Each entry in the map should correspond to a row in a junction
table between models and scripts (note that this is done elsewhere)

- The `ML.loadModel` primitive has been changed to take a `modelName`
which are provided to the script evaluator. This is similar to how all
of the bridge primitives work (i.e. only an identifier is provided
directly by the user, which is resolved by the script evaluator). The
old behavior is still available under `ML.unsafeLoadScript`, which just
takes a filepath. The advantage of this is that `loadModel` can only be
used with names known to be valid -- the correct filepath is derived for
the user