[inferno-vc] Cached client and server logging. More granular cache locks. Cleanup (#143)
This PR adds logging to cached client to see hits/misses and makes the
thundering-herd-prevention locks more granular while only taking the
lock when actually fetching upstream. Also does some cleanup in that
module so the thundering-herd-prevention logic is easier to inspect and
functions have names that better describe what they do. This is 3rd take
of #139 but this time the semantics of fetchVCObjectClosureHashes is not
modified