[inferno-ml] Change parameter input/output representation (#147)
The previous way that parameter inputs/outputs were stored, with both
condensed into a single `Map Ident ...`, turned out to be quite
nightmarish to work with. We are going to be doing things like filtering
or querying parameters by output, so we need it to be easier to look
outputs up. So this changes the inputs/outputs back to the way they used
to be, e.g. two separate `Map`s