[inferno-ml] Use connection pool (#149)
Currently, the entire server runs within `withConnect` and holds a
single `Connection` within its `Env`. If Postgres is unreachable for
some reason when the server starts, the process will hang until it's
killed by an exception (which takes quite some time). If the connection
fails while the server is running, the entire process will be brought
down. This switches the server to use a `Pool` from `resource-pool`
instead to create/hold connections on demand