[inferno-ml] Move `inferno-ml-server` to system service (#152)
Currently `inferno-ml-server.service` runs as a systemd user service.
However, this doesn't actually work when testing the EC2 instances, as a
shell must be opened for the `inferno` user before the service starts.
We need it to start immediately after init

I've tested this with a real EC2 built using a testing image config and
it fixes the problem

NOTE: We should eventually create an `inferno-ml-server` system user to
run the service. However, for debugging and permissions purposes, it's
easier to continue to run it as `inferno` for now. I've made an issue to
fix this: https://github.com/plow-technologies/inferno/issues/151. But
this fix is needed urgently and I don't want to hold things up fixing
that.