[inferno-ml] Make sure evaluation environment is always valid (#153)
Currently, if part of the `EvaluationEnvironment` is missing, e.g.
inputs or outputs, script evaluation will fail. In this case, the most
sensible thing to do is just use whatever is linked to the param. This
allows for selectively overriding certain parts of the environment
without the whole thing failing