[inferno-ml] Fix `toDevice` primitive (#162)
Currently, `toDevice` takes a user-supplied `Text` as the name of the
device to move the tensor. This is not ideal since it's easy to make a
mistake and get some nasty Torch error. This changes `toDevice` to take
a `{#cpu,#cuda}` enum as an argument and renames the existing primitive
to `toDeviceUnsafe`