[inferno-ml] Don't `error` when moving tensors (#167)
We currently have a problem where the "pure" `toDevice` throws an
`error` when a tensor can't be moved.

This fixes the `toDevice` primitive used by `inferno-ml-server` to
handle that case. We discussed this and decided not to throw a
synchronous exception when this happens. Instead, we can log the
failure.

This has the advantage of allowing `toDevice` to be run without script
modification on different `inferno-ml-server` instances (i.e. with or
without CUDA).

We do need to at least log the failure though, so I've made `mlPrelude`
parameterized by a `toDevice` implementation -- we can't properly log in
`inferno-ml`, so the implementation has been moved to
`inferno-ml-server`.