[inferno-ml] Swap order of elements in `makeWrites` (#164)
We've decided to swap the order of `makeWrites` so it takes a `array of
('a, t)` instead of `array of (t, 'a)` as it currently does. This is to
make it consistent with other things that take pairs of values and
times.

This is working as far as `inferno-ml-server` goes (integration tests
pass), but we need to do some updates elsewhere before merging this.