[inferno-ml] Fix `forward` primitive (#171)
Similar to other issues we've had in the past, Hasktorch's `forward` is
presented as a "pure" function but actually uses `unsafePerformIO`
internally. This leads to use not catching `CppException`s in case the
Torch interpreter runs into a problem. This fixes the primitive by using
`evaluate` to lift Hasktorch's `forward` into `IO` and then
catching/rethrowing the exception as an Inferno `EvalError`.
`inferno-ml-server` should then send a proper error response in case of
an exception.

I've run this against both sets of Inferno ML integration tests that we
have an it works with existing (good) scripts. I also tried running some
known-bad Torchscript `ScriptModules` and the exception is caught.