[inferno-ml] Store model version size (#172)
Since the `contents` of a `ModelVersion` are immutable, we can calculate
its size once and store it in the DB. Currently we only need the size
once in `inferno-ml-server` (when caching a model version) but now we
want to display them elsewhere. Calculating it each time is wasteful.

This also includes a few other changes:
- `SqlError`s are now caught and rethrown so they show up as error
traces and in error responses
- The `unique` constraint on `models` has been changed so that a new
model can be created with the same name as (a) previously terminated
model(s)