[inferno-ml] Add logging primitives (#174)
This adds the ability to log values to a "console" directly from
scripts. This is more useful for Inferno ML because debugging can be
difficult with the different types involved.

- Adds `Print` module to Inferno ML prelude
- Adds implementation for `inferno-ml-server`

NOTE: Previously I tried to get wildcard bindings working (e.g. `let _ =
Print.print... in`) but it requires more out-of-scope work