[inferno-ml] Add `inferno-ml-compat` to generate `ML` module (#177)
Currently we have two places (here in `inferno` and elsewhere) where we
need to use the `ML` module primitive definitions. The current
`inferno-ml` package has a hard dependency on `hasktorch`, however,
which we can't use in the other place we need the `ML` module, so we've
had to duplicate the module definition and keep the two in sync
manually.

This fixes that by adding an `inferno-ml-compat` package that, given
certain types (tensor, script module, etc...), will generate the `ML`
module from the primitives contained in a group of record-of-functions.

The `inferno-ml` package instantiates this using `Tensor`,
`ScriptModule`, etc... from Hasktorch, and provides primitive
implementations that can be evaluated at runtime. Elsewhere we can use
dummy types for type-checking and LSP purposes.

This keeps the `ML` module in sync but avoids a direct dependency on
Hasktorch where we can't use it.