[inferno-ml] Run `inferno-ml-server` within systemd OOM policy (#180)
- Adds an `OOMPolicy` for the `inferno-ml-server` process with
`MemoryMax` as a percentage of system memory
- Adds warning logs if the server is restarting from an OOM event
(helpful for debugging)
- Adds `configuration.memoryMax` to the NixOS config for the server with
corresponding adjustments on the Haskell type

This is needed because scripts can sometimes consume all system memory
until kernel memory reclamation begins and the entire system slows to a
crawl and freezes from a user's point of view. It's better to just kill
the process and restart it in this case, otherwise we can't even e.g.
poll the process status. A subsequent PR will add in-process memory
monitoring to preempt the OOM and just kill the thread that's evaluating
the script, but that's a different task (and requires the cgroup
information that's created by introducing the `OOMPolicy` to the systemd
service).

Example log after triggering OOM:

```
inferno-ml-server[1143]: 2025-09-12 08:41:30.863259853 UTC: [WARN] Server is restarting from out-of-memory event triggered at 2025-09-12 08:41:25 UTC
```