[inferno-ml] Add in-process memory monitor to preempt systemd OOM (#183)
Adds an in-process monitor for memory usage during script evaluation by
reading the cgroup memory information for the `inferno-ml-server`
systemd service. If we get close to the OOM threshold, the server will
respond with an error message and then restart itself. This avoids
having script evaluation interrupted **without** an error message, which
is the case currently. Since we have `Restart=always` defined for the
systemd service, the process memory is freed and the service restarts.

At first I wanted to avoid having the restart, but despite my best
efforts I could not get GHC or libc arenas (from Torch FFI) to free
enough memory. That would leave us in a state where memory was still
allocated but not freed, and any subsequent script evaluations would
fail. We would still get the error message, but the server would need to
be manually restarted to free all of the memory. So instead a
self-restart is scheduled after Warp has enough time to serve the
request with the error details.

Another alternative would be to run script evaluation within a separate
process within systemd constraints, but that would involve rewriting a
significant portion of `inferno-ml-server` and we don't have the time
for that.

Now if we are close enough to the OOM threshold we'll see error messages
like this:

```
"Memory usage of script evaluator has exceeded limit, script evaluation memory usage: 0.07 GB"
``` 

(note that I had to intentionally lower the OOM threshold to 7% of
system memory to get the error message, a bit contrived but illustrates
the principle)