[inferno-ml] Add per-server configuration service (#185)
We've realized that we need a way to configure Inferno ML server
instances on a per-instance basis. If we include everything in the NixOS
configuration, then we will end up building per-instance images, which
we want to avoid.

So this adds `inferno-ml-configure`, which is a small service that runs
before `inferno-ml-server` and writes per-instance configuration for the
main server.

At the moment, this is mostly to fix issues with not getting logs from
instances, but we've wanted a way to do this for a while so now seems to
be a good time to add it.