[inferno-ml-server] Fix argument position when applying scripts (#188)
He had an issue where `inferno-ml-server` expected function arguments to
be in a specific order, but we can't always expect that (as a subtle bug
recently demonstrated). So this fixes it to extract all arguments from
the `Lam` in the `VCFunction`, then do a `Map.lookup` on all
inputs/outputs/models to reconstruct the correct ordering to apply the
function. This makes `inferno-ml-server` work with any argument
ordering, which is far more resilient and doesn't rely on fragile
ordering/sorting guarantees to be synchronized everywhere